Extracted interfaces from userUtils to common interface
